==> code/03_alignment_analysis/reddit/01_load_brands.ipynb <==
## Load Brand Data 
The script here compiles a list of skin care brand from various kaggle datasets. The resulting list will be used to filter out Reddit comments referencing unrelated brands for a more accurate brand alignment analysis. 

import pandas as pd
import praw
from bs4 import BeautifulSoup # library for parsing html
import os 

data_path = '../../../data/brand/'

 # cosmetic company tend to also sell skincare, so using this to build list of as many skin care brand as possible
 # these data sets are found from kaggle 
 
cosmetics_list = pd.read_csv(os.path.join(data_path, "cosmetics.csv"))
product_info_data = pd.read_csv(os.path.join(data_path, "product_info.csv"))
skincare_list = pd.read_csv(os.path.join(data_path, "skincare.csv"))
skincare2_list = pd.read_csv(os.path.join(data_path, "skincare2.csv"))
all_brands = list(set(cosmetics_list["Brand"].str.lower().unique().tolist() + product_info_data["brand_name"].str.lower().unique().tolist() + skincare_list["Brand"].str.lower().unique().tolist() + skincare2_list["brand"].str.lower().unique().tolist()))
pd.DataFrame(all_brands, columns=["brand_name"]).sort_values(by="brand_name").to_csv(data_path + "all_brands.csv", index=False)
