==> code/03_alignment_analysis/reddit/03_clean_reddit_comment.ipynb <==
# Comment text cleaning 
1. Break each comment down into as a individual row 
2. Preprocess each row
3. Export cleaned comment data 


import pandas as pd
import os

import re
import ast

import nltk
from nltk import pos_tag
from nltk.tokenize import word_tokenize
from nltk.stem.snowball import SnowballStemmer
from nltk.corpus import stopwords as nltk_stopwords
output_path = "../../../data/alignment_analysis/"
os.makedirs(output_path, exist_ok=True)


### Load Subreddit data

subreddit_df = pd.read_csv(output_path + "subreddit_data.csv")
subreddit_df.head()

all_brand_df = pd.read_csv(f"../../../data/all_brands.csv")
all_brands = set(all_brand_df["brand_name"].dropna().str.lower().unique())

### Break each comment into a row

subreddit_df["top_comments"] = subreddit_df["top_comments"].apply(
    lambda x: ast.literal_eval(x) if isinstance(x, str) else x
)
def mentions_other_brand(comment, current_brand):
    comment = str(comment).lower()
    current_brand = str(current_brand).lower()
    for brand in all_brands:
        if brand != current_brand and brand in comment:
            return True
    return False

rows = []
for _, row in subreddit_df.iterrows():
    for comment in row["top_comments"]:
        rows.append({
            "post_id": row["post_id"],
            "title": row["title"],
            "brand": row["brand"],
            "subreddit_name": row["subreddit_name"],
            "comment": comment
        })
        
comments_df = pd.DataFrame(rows)
print(comments_df.shape)
# filter out comments that mention other brands
comments_df = comments_df[~comments_df["comment"].apply(lambda x: mentions_other_brand(x, comments_df["brand"]))]
print(comments_df.shape)
comments_df.head()



### Preprocess comment

custom_stopwords = [
    # Brand names (removed from analysis)
    'estee', 'lauder', 'tarte', 'fenty', 'glossier', 'cosrx', 'etude',
    'sulwhasoo', 'laneige', 'innisfree', 'elf',

    # Platform-related
    'video', 'youtube', 'tiktok', 'instagram', 'reel', 'feed',
    'post', 'stories', 'caption', 'social', 'media',

    # Engagement / action words
    'like', 'likes', 'comment', 'comments', 'share', 'save', 'follow', 'subscribe',
    'tag', 'click', 'link', 'bio', 'visit', 'dm', 'available', 'check',

    # Time / filler
    'today', 'now', 'new', 'soon', 'launch', 'launching', 'stay', 'tune', 'coming', 'back',

    # General beauty-related terms
    'beauty', 'skin', 'skincare', 'routine', 'makeup', 'product', 'products',
    'face', 'body', 'glow', 'look', 'formula', 'texture', 'result',

    # Emoji / symbols
    '✨', '🔥', '💧', '💫', '😍', '💖', '🌟', '💥', '🧴', '📦', '🛍️',

    # Overused positive adjectives
    'feel', 'love', 'use', 'try', 'amazing', 'favorite', 'best', 'perfect', 'must', 'obsessed',

    # Promotional terms
    'shop', 'buy', 'discount', 'deal', 'sale', 'off', 'gift', 'giveaway', 'free', 'offer',

    # Conversation filler
    'hey', 'hello', 'welcome', 'thank', 'you', 'everyone', 'guys', 'hi', 'omg', 'pls', 'yay', 'get', 'got', 'let', 'us',
    "follow", "like", "likes", "just", "really", "thanks", "omg",
    "hi", "hey", "pls", "please", "gonna", "tbh", "honestly",
    "lol", "lmao", "idk", "link", "click", "watch", "dm", "recommend",
]

stopword_set = set(nltk_stopwords.words('english')).union(custom_stopwords)

def preprocess_comment(comment):
    if not isinstance(comment, str):
        return ""
    # lower case 
    comment= comment.lower()
    comment = re.sub(r"http\S+|www\S+|@\w+|#\w+|u/\w+", "", comment)
    comment = re.sub(r"[^a-z\s]", " ",comment)
    comment = re.sub(r"\s+", " ", comment).strip()
    tokens = word_tokenize(comment)
    
    # remove stopwords 
    filtered_tokens = [t for t in tokens if t.isalpha() and t not in stopword_set and len(t) > 2] 
    return " ".join(filtered_tokens)

comments_df["cleaned_comment"] = comments_df["comment"].apply(preprocess_comment)

comments_df.to_csv(output_path + "subreddit_comment_data.csv", index=False)
comments_df.head()
comments_df.shape
